+++ hub/core/compression/__init__.py
@@ -1,5 +1,6 @@
 import sys
 import struct
+import re
 
 
 if sys.byteorder == "little":
@@ -13,7 +14,13 @@ STRUCT_HHB = struct.Struct(">HHB")
 STRUCT_II = struct.Struct(">ii")
 
 
+def re_find_first(pattern, string):
+    for match in re.finditer(pattern, string):
+        return match
+
+
 from .base_compressor import BaseCompressor
 from .image.jpeg import JPEG
 from .image.png import PNG
 from .audio.mp3 import MP3
+
